Added comments outlining whilebankopencook fix.
-Outlined the steps needed to fix: When a player clicks "cook" while
bank interface is open, because we only check the players gear once the
interface is closed, we do not have updated global booleans when we need
them.

Perhaps, I should have a class encapsulating state, and when
.getInventoryState is called, it fetches the players inventory ensuring
that we only call it when needed, and that the check, and warn / menu
option clicked has the most updated information.